+++ src/main/java/com/boardgo/domain/review/service/ReviewCommandUseCase.java
@@ -1,0 +1,7 @@
+package com.boardgo.domain.review.service;
+
+import com.boardgo.domain.review.controller.request.ReviewCreateRequest;
+
+public interface ReviewCommandUseCase {
+    void create(ReviewCreateRequest reviewType, Long reviewerId);
+}
