+++ src/main/java/com/boardgo/domain/review/service/facade/ReviewCommandFacade.java
@@ -1,0 +1,7 @@
+package com.boardgo.domain.review.service.facade;
+
+import com.boardgo.domain.review.controller.request.ReviewCreateRequest;
+
+public interface ReviewCommandFacade {
+    void create(ReviewCreateRequest reviewType, Long reviewerId);
+}
